[Feature] Provide `log.check()` to log output based on boolean tests
It is currently possible to create debugging output based on the
results of a boolean condition by writing code like so:

    if enemy.hp <= 0 then
        log.warn("Spawned an enemy with no health.")
    end

This patch provides a new function, `log.check()`, allowing us to
rewrite the above as this:

    log.check(enemy.hp <= 0, "Spawned an enemy with no health.")

We are hardly saving much with regard to lines-of-code.  However,
`log.check()` lets us omit the second parameter, meaning we can write:

    log.check(enemy.hp <= 0)

This will produce a warning, via `log.warn()`, with the message
`"Check failed"`.  Thus `log.check()` shares the same order and
optional parameters as Lua's standard `assert()`.  The key difference
is that `log.check()` will not raise an error like `assert()`.  It
only performs logging; if the failure of a condition is so serious
that it requires raising an error then we must still resort to
functions like `assert()` and `error()`.

By default `log.check()` uses the `"warn"` level.  It accepts an
optional third parameter: the name of a level at which to produce the
output.  For example:

    log.check(player.xp >= MAX_FOR_CLASS,
              "Player exceeded experience points for class.",
              "trace")

This will log the message by using the `log.trace()` function.

The naive implementation of `log.check()` looks like this:

    log.check = function (value, message, level)
        local level = level or "warn"
        if not value then
            log[level](message)
        end
        return value
    end

This is a flawed definition because all messages logged will use the
filename and line number where we define `log.check()`.  What we want,
however, is for the function to display the file information regarding
the call-site of `log.check()`.  We can call debug.getinfo() from
within `log.check()` to get the correct information we wish to
display, but we have no way of sending that to the log output.

Our solution for this is to refactor a large chunk of the logic within
the six base logging functions.  This patch introduces the new,
private `create_output()` function for that purpose.  It accepts not
only the level of message to create, e.g. `"error"`, but also takes
file information that the original logic obtained from
`debug.getinfo()`.  Refactoring the logic into `create_output()`
allows us to obtain the correct `debug.getinfo()` information from
within `log.check()`, ultimately making it possible for the function
to display data about the call-site of `log.check()` just like all
other log functions.